all: move marshal to stdlib/marshal
