+++ stdlib/marshal/marshal.go
